Fix #1473:  Add navigation and create feeds page (#1475)
* add navigation and create feeds page

* separated callbacks to be page specific

* separated views by page

* write test for navigation
